:hammer: refactor: Seperate navigation from header (#7)
